+++ packages/celest_db_studio/lib/src/driver.web.dart
@@ -1,0 +1,7 @@
+import 'package:celest_db_studio/src/driver.dart';
+
+abstract final class DriverImpl {
+  static Future<Driver> connect(Uri uri, {String? authToken}) async {
+    throw UnimplementedError('DriverImpl.connect() is not implemented on Web');
+  }
+}